project done
I originally downloaded a copy of our final RUfit app from the project's github and began work with that. My additions are the Leader class, LeaderboardActivity, its layout, and the layouts of the leaderboard dialog and the facebook sharing dialog. I added code to FinishRunActivity, DetailWorkout, StatisticsActivity, and their associated layouts as well.